requirements.txt
alembic==1.0.11
Click==7.0
Flask==1.0.3
Flask-Assets==0.12
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.13
MarkupSafe==1.1.1
psycopg2==2.8.3
py-bcrypt==0.4
pyScss==1.3.5
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.5
webassets==0.12.1
Werkzeug==0.15.4
WTForms==2.2.1
